# server/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Server for Legacy App",
        "title": "Legacy API",
        "contact": {
            "name": "David Oduneye and Akshay Dupuguntla",
            "email": "oduneye.d@northeastern.edu and dupuguntla@northeastern.edu"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/profiles/": {
            "get": {
                "description": "Returns all profiles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Gets all profiles",
                "operationId": "get-all-profiles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Profile"
                            }
                        }
                    },
                    "404": {
                        "description": "Failed to fetch profiles",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Creates a profile",
                "operationId": "create-profile",
                "parameters": [
                    {
                        "description": "Profile",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.ProfileDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "400": {
                        "description": "Failed to create profile",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/profiles/complete/{pid}": {
            "patch": {
                "description": "Sets onboarding complete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Sets onboarding complete",
                "operationId": "set-onboarding-complete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProfileID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "400": {
                        "description": "Failed to set onboarding complete",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/profiles/response/{pid}": {
            "patch": {
                "description": "Inserts an onboarding response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Inserts an onboarding response",
                "operationId": "insert-onboarding-response",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProfileID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "OnboardingResponse",
                        "name": "onboardingResponse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "400": {
                        "description": "Failed to insert onboarding response",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/profiles/{pid}": {
            "get": {
                "description": "Returns a profile from profile id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Gets a profile from profile id",
                "operationId": "get-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProfileID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "404": {
                        "description": "Failed to fetch profile",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Deletes a profile",
                "operationId": "delete-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProfileID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted profile",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Failed to delete profile",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Updates a profile",
                "operationId": "update-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProfileID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Profile",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.ProfileDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "400": {
                        "description": "Failed to update profile",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docmodels.ProfileDTO": {
            "type": "object",
            "properties": {
                "completed_onboarding_response": {
                    "type": "boolean"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "onboarding_response": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/docmodels.UserDTO"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "docmodels.UserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firebase_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "persona_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.Profile": {
            "type": "object",
            "required": [
                "date_of_birth",
                "name",
                "phone_number",
                "user_id"
            ],
            "properties": {
                "completed_onboarding_response": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string"
                },
                "onboarding_response": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}